requirements.txt
# Core testing frameworks
pytest>=7.4.4
pytest-bdd>=8.1.0
pytest-html>=4.1.1
pytest-json-report>=1.5.0
pytest-xdist>=3.5.0
pytest-repeat>=0.9.3
pytest-timeout>=2.2.0
pytest-mock>=3.12.0

# API testing
requests>=2.31.0
jsonschema>=4.20.0
pydantic>=2.6
pydantic-settings>=2.0
httpx>=0.25.2

# UI testing
playwright>=1.48.0

# Mobile Testing
Appium-Python-Client>=3.1.1
selenium>=4.15.2

# Performance testing
locust>=2.17.0
pytest-benchmark>=4.0.0

# Data handling
faker>=20.1.0
pandas>=2.2.3,<3
openpyxl>=3.1.2

# Utilities
python-dotenv>=1.1.1
allure-pytest-bdd>=2.15.0
jinja2>=3.1.2
pyyaml>=6.0.1
weasyprint>=66.0

# Logging and monitoring
structlog>=23.2.0
colorlog>=6.8.0
loguru>=0.7.2